+++ app/gui/src/dashboard/components/AriaComponents/Checkbox/CheckboxGroup.tsx
@@ -1,8 +1,4 @@
-/**
- * @file
- *
- * A CheckboxGroup allows users to select one or more items from a list of choices.
- */
+/** @file A selector for one or more items from a list of choices. */
 import type { CheckboxGroupProps as AriaCheckboxGroupProps } from '#/components/aria'
 import { CheckboxGroup as AriaCheckboxGroup, mergeProps } from '#/components/aria'
 import { mergeRefs } from '#/utilities/mergeRefs'
@@ -10,15 +6,17 @@ import { omit } from '#/utilities/object'
 import { forwardRef } from '#/utilities/react'
 import type { VariantProps } from '#/utilities/tailwindVariants'
 import { tv } from '#/utilities/tailwindVariants'
-import type { CSSProperties, ForwardedRef, ReactElement } from 'react'
+import type { CSSProperties, ForwardedRef, ReactElement, ReactNode } from 'react'
 import type { FieldVariantProps } from '../Form'
 import { Form, type FieldPath, type FieldProps, type FieldStateProps, type TSchema } from '../Form'
 import type { TestIdProps } from '../types'
 import { CheckboxGroupProvider } from './CheckboxContext'
 
-/** Props for the {@link CheckboxGroupProps} component. */
-export interface CheckboxGroupProps<Schema extends TSchema, TFieldName extends FieldPath<Schema>>
-  extends FieldStateProps<AriaCheckboxGroupProps, Schema, TFieldName>,
+/** Props for the {@link CheckboxGroup} component. */
+export interface CheckboxGroupProps<
+  Schema extends TSchema,
+  TFieldName extends FieldPath<Schema, readonly string[]>,
+> extends FieldStateProps<AriaCheckboxGroupProps, Schema, TFieldName, readonly string[]>,
     FieldProps,
     FieldVariantProps,
     Omit<VariantProps<typeof CHECKBOX_GROUP_STYLES>, 'disabled' | 'invalid'>,
@@ -26,7 +24,7 @@ export interface CheckboxGroupProps<Schema extends TSchema, TFieldName extends F
   readonly className?: string
   readonly style?: CSSProperties
   readonly checkboxRef?: ForwardedRef<HTMLInputElement>
-  readonly children: ReactElement | ((props: AriaCheckboxGroupProps) => ReactElement)
+  readonly children: ReactNode | ((props: AriaCheckboxGroupProps) => ReactNode)
 }
 
 const CHECKBOX_GROUP_STYLES = tv({
@@ -34,11 +32,11 @@ const CHECKBOX_GROUP_STYLES = tv({
   variants: { fullWidth: { true: 'w-full' } },
 })
 
-/** A CheckboxGroup allows users to select one or more items from a list of choices. */
+/** A selector for one or more items from a list of choices. */
 export const CheckboxGroup = forwardRef(
-  <Schema extends TSchema, TFieldName extends FieldPath<Schema>>(
+  <Schema extends TSchema, TFieldName extends FieldPath<Schema, readonly string[]>>(
     props: CheckboxGroupProps<Schema, TFieldName>,
-    ref: ForwardedRef<HTMLFieldSetElement>,
+    ref: ForwardedRef<HTMLDivElement>,
   ): ReactElement => {
     const {
       children,
@@ -48,7 +46,7 @@ export const CheckboxGroup = forwardRef(
       defaultValue: defaultValueOverride,
       isDisabled = false,
       isRequired = false,
-      isInvalid = false,
+      isInvalid,
       isReadOnly = false,
       label,
       name,
@@ -70,6 +68,7 @@ export const CheckboxGroup = forwardRef(
         {...(defaultValueOverride != null && { defaultValue: defaultValueOverride })}
         render={({ field, fieldState }) => {
           const defaultValue = defaultValueOverride ?? formInstance.control._defaultValues[name]
+          const invalid = isInvalid ?? fieldState.invalid
           return (
             <>
               <CheckboxGroupProvider
@@ -84,7 +83,7 @@ export const CheckboxGroup = forwardRef(
                 <AriaCheckboxGroup
                   {...mergeProps<AriaCheckboxGroupProps>()(omit(checkboxGroupProps, 'validate'), {
                     className: styles,
-                    isInvalid,
+                    isInvalid: invalid,
                     isDisabled,
                     isReadOnly,
                     name,
@@ -101,7 +100,7 @@ export const CheckboxGroup = forwardRef(
                       description={description}
                       isRequired={isRequired}
                       fullWidth={fullWidth}
-                      isInvalid={isInvalid || fieldState.invalid}
+                      isInvalid={invalid}
                       variants={fieldVariants}
                       {...checkboxGroupProps}
                     >