+++ app/gui/src/dashboard/components/AriaComponents/Dialog/Dialog.tsx
@@ -12,9 +12,10 @@ import * as suspense from '#/components/Suspense'
 
 import * as mergeRefs from '#/utilities/mergeRefs'
 
+import { DialogDismiss } from '#/components/AriaComponents'
 import { useEventCallback } from '#/hooks/eventCallbackHooks'
 import { useMeasure } from '#/hooks/measureHooks'
-import { motion, type Spring } from '#/utilities/motion'
+import { LayoutGroup, motion, type Spring } from '#/utilities/motion'
 import type { VariantProps } from '#/utilities/tailwindVariants'
 import { tv } from '#/utilities/tailwindVariants'
 import { Close } from './Close'
@@ -321,7 +322,7 @@ function DialogContent(props: DialogContentProps) {
   }
 
   return (
-    <>
+    <LayoutGroup>
       <MotionDialog
         layout
         transition={TRANSITION}
@@ -400,7 +401,7 @@ function DialogContent(props: DialogContentProps) {
       </MotionDialog>
 
       <dialogStackProvider.DialogStackRegistrar id={dialogId} type={TYPE_TO_DIALOG_TYPE[type]} />
-    </>
+    </LayoutGroup>
   )
 }
 
@@ -551,4 +552,5 @@ const DialogHeader = React.memo(function DialogHeader(props: DialogHeaderProps) 
 })
 
 Dialog.Close = Close
+Dialog.Dismiss = DialogDismiss
 Dialog.Trigger = DialogTrigger